requirements.txt
customtkinter>=5.2.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
nuitka>=2.0.0       # packaging only — not needed to run the app
